small structure improvements
