+++ src/lib/emoteList.ts
@@ -1,0 +1,12 @@
+import { type Emote } from "~/types/GameTypes";
+
+const emoteList: Array<Emote> = [
+  { slug: "cat", name: "Pouty Cat", pathName: "pout-cat.png"},
+  { slug: "bunny", name: "Hurry Up Bunny", pathName: "bunny.png"},
+  { slug: "chicken", name: "Thumbs Up Chicken", pathName: "chicken.png"},
+  { slug: "koda", name: "Koda, Hi!", pathName: "dog.png"},
+  { slug: "ostrich", name: "Angry Ostrich", pathName: "ostrich.png"},
+  { slug: "parrot", name: "Copy Parrot", pathName: "parrot.png"},
+];
+
+export default emoteList;
