+++ src/styles/globals.css
@@ -78,7 +78,6 @@ body.signed-in::before {
   }
 }
 
-
 @keyframes ping {
   25% {
       transform: scale(1.3);
